youtube-shorts-maker/pyproject.toml
[project]
name = "youtube-shorts-maker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.21.0",
    "google-genai>=1.57.0",
    "python-dotenv>=1.2.1",
]
